requirements.txt
streamlit
python-docx
PyPDF2
faiss-cpu
torch
transformers
google-generativeai
python-dotenv